Move win32 specific stuff to win32/findfile.c
